# docs/partner-api.openapi.yaml
openapi: 3.1.0
info:
  title: Pallares Partner Lead API
  version: v1
  description: |
    Cursor-sync Partner API for Pallares.us lead ingestion and structured feedback.

    **Primary auth:** `x-api-key: ppl_...`  
    **Also accepted:** `Authorization: Bearer ppl_...`

    Leads are served exclusively from the `partner_leads_v1` view. A lead is eligible only when:
    - `enrichment_status = enriched` and `enriched_json` is present
    - `confidence` is not `Low`
    - `lead_score >= 25` (current `min_export_score`)
    - `is_verified_decision_maker(...)` is true — requires `verification_level = verified`
      plus a grounded, non-placeholder name + decision-maker role + local non-toll-free phone

    Partial verification and Google mainline-only phones are **not** partner-eligible.
    `primary_phone` is a locally callable `best_contact_phone` only (placeholders
    such as `Not found` are null; never coalesced from Google `main_phone`).
    Path `place_id` values may contain slashes (e.g. `places/ChIJ...`); URL-encode them
    (`places%2FChIJ...`) or pass the unencoded multi-segment path — both are accepted.

    **Idempotency:** POST `/leads/{place_id}/outcome`, `/leads/{place_id}/touches`, and
    `/feedback/batch` accept an optional `Idempotency-Key` header. Successful responses
    are stored per partner key and returned on replay.

    ## Roadmap

    `POST /orders` and `POST /webhooks` are OpenAPI stubs only (`x-stability: experimental`).
...
servers:
...
paths:
...
  /leads/{place_id}:
    get:
      summary: Lead detail
      parameters:
        - $ref: "#/components/parameters/PlaceId"
      responses:
        "200":
          description: Eligible lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Preferred partner authentication header (`ppl_...`)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ppl_
      description: Compatibility alternative to x-api-key
  parameters:
    PlaceId:
      in: path
      name: place_id
      required: true
      schema:
        type: string
      description: |
        Lead place id. May contain slashes (Google Places resource name). Prefer
        URL-encoding (`encodeURIComponent`) when building clients.
...
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    LeadListItem:
      type: object
      properties:
        lead_id:
          type: string
        place_id:
          type: string
        lead_type:
          type: string
          enum: [client, vendor]
        business_name:
          type: string
        category_key:
          type: string
          nullable: true
        market_key:
          type: string
          nullable: true
        primary_phone:
          type: string
          nullable: true
          description: >
            Atomic DM phone (`best_contact_phone` only when locally callable).
            Placeholder sentinels such as `Not found` are null. Never coalesced
            from Google `main_phone`.
        verification_level:
          type: string
          enum: [verified]
          description: Partner-eligible leads are verified only (partial excluded)
        updated_at:
          type: string
          format: date-time
...
    LeadDetailResponse:
      type: object
      properties:
        data:
          allOf:
            - $ref: "#/components/schemas/LeadListItem"
            - type: object
              properties:
                site_contacts:
                  type: array
                  items:
                    type: object
                evidence_urls:
                  type: array
                  items:
                    type: string
                fact_summaries:
                  type: object
                score_breakdown:
                  type: object
                coordinates:
                  type: object
                  properties:
                    latitude:
                      type: number
                      nullable: true
                    longitude:
                      type: number
                      nullable: true
        meta:
          type: object
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    NotFound:
      description: Lead not found or not eligible
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    RateLimited:
      description: Rate or daily row budget exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
security:
  - apiKeyAuth: []
  - bearerAuth: []